refactor: reorganize and restructure tabs and divider components
Moved tabs and divider components to a new folder structure to improve organization and maintainability. Updated export paths accordingly to align with the new structure.
