+++ src/components/navigation/tabs/tab/index.ts
@@ -1,0 +1,3 @@
+export * from './tab.component';
+export * from './tab.interface';
+export * from './tab.style';
