+++ src/index.ts
@@ -5,9 +5,9 @@ export * from './components/communication/progress-indicator';
 export * from './components/communication/snackbar';
 export * from './components/containment/card';
 export * from './components/containment/carousel';
-export * from './divider';
+export * from './components/containment/divider';
+export * from './components/navigation/tabs';
 export * from './icon';
-export * from './navigation/tabs';
 export * from './selection/slider';
 export * from './selection/switch';
 export * from './text-inputs';
