--- src/navigation/tabs/index.ts
@@ -1,3 +1,0 @@
-export * from './Tab';
-export * from './Tabs';
-export * from './TabStyle';
